Fix select next and prev in list, do not wrap around
Consider that next to a single element, is nothing. This is important
because that function is used to set the next `selected` before
removing an element, and we cannot set `selected` to an element that
was just removed.

The UI function selecting clients still needs to wrap around, but that
is done outside of the linked list.